refactor: 색칠 함수 분리
