+++ src/components/diary/Canvas.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import DrawImage from "@/lib/DrawImage";
-import { convertHexToRgba, getPixelColor, isSameColor, setPixel } from "@/lib/Paint";
+import { convertHexToRgba, floodFill } from "@/lib/Paint";
 import Redo from "@/lib/Redo";
 import ReDraw from "@/lib/ReDraw";
 import SetCanvasContext from "@/lib/SetCanvasContext";
@@ -117,40 +117,11 @@ const Canvas = ({
     setPathStep(pathStep + 1);
   };
 
-  const floodFill = (x: number, y: number, fillColor: Uint8ClampedArray) => {
-    const imageData = ctx?.getImageData(0, 0, ctx.canvas.width, ctx.canvas.height);
-    if (imageData) {
-      const visited = new Uint8Array([imageData.width, imageData.height]);
-      const targetColor = getPixelColor(imageData, x, y) as Uint8ClampedArray;
-
-      if (!isSameColor(targetColor, fillColor)) {
-        const stack = [{ x, y }];
-        while (stack.length > 0) {
-          const child = stack.pop();
-          if (!child) return;
-          const currentColor = getPixelColor(imageData, child.x, child.y);
-          if (
-            !visited[child.y * imageData.width + child.x] &&
-            isSameColor(currentColor as Uint8ClampedArray, targetColor)
-          ) {
-            setPixel(imageData, child.x, child.y, fillColor);
-            visited[child.y * imageData.width + child.x] = 1;
-            stack.push({ x: child.x + 1, y: child.y });
-            stack.push({ x: child.x - 1, y: child.y });
-            stack.push({ x: child.x, y: child.y + 1 });
-            stack.push({ x: child.x, y: child.y - 1 });
-          }
-        }
-        ctx?.putImageData(imageData, 0, 0);
-      }
-    }
-  };
-
   const paintCanvas = (e: React.MouseEvent<HTMLCanvasElement>) => {
     const curPos = { x: e.nativeEvent.offsetX, y: e.nativeEvent.offsetY };
     if (!curPos) return;
     const currentColor = convertHexToRgba(lineCustom.lineColor);
-    floodFill(curPos.x, curPos.y, currentColor);
+    floodFill(curPos.x, curPos.y, currentColor, ctx);
   };
 
   return (
