+++ src/lib/Paint.tsx
@@ -38,3 +38,32 @@ export const setPixel = (imageData: ImageData, x: number, y: number, color: Uint
   imageData.data[offset + 2] = color[2];
   imageData.data[offset + 3] = color[3];
 };
+
+export const floodFill = (x: number, y: number, fillColor: Uint8ClampedArray, ctx: CanvasRenderingContext2D | null) => {
+  const imageData = ctx?.getImageData(0, 0, ctx.canvas.width, ctx.canvas.height);
+  if (imageData) {
+    const visited = new Uint8Array([imageData.width, imageData.height]);
+    const targetColor = getPixelColor(imageData, x, y) as Uint8ClampedArray;
+
+    if (!isSameColor(targetColor, fillColor)) {
+      const stack = [{ x, y }];
+      while (stack.length > 0) {
+        const child = stack.pop();
+        if (!child) return;
+        const currentColor = getPixelColor(imageData, child.x, child.y);
+        if (
+          !visited[child.y * imageData.width + child.x] &&
+          isSameColor(currentColor as Uint8ClampedArray, targetColor)
+        ) {
+          setPixel(imageData, child.x, child.y, fillColor);
+          visited[child.y * imageData.width + child.x] = 1;
+          stack.push({ x: child.x + 1, y: child.y });
+          stack.push({ x: child.x - 1, y: child.y });
+          stack.push({ x: child.x, y: child.y + 1 });
+          stack.push({ x: child.x, y: child.y - 1 });
+        }
+      }
+      ctx?.putImageData(imageData, 0, 0);
+    }
+  }
+};
