+++ src/lib/Undo.tsx
@@ -1,19 +1,10 @@
 import { ChangedoProps } from "@/types/Canvas";
 import GetRatio from "./GetRatio";
 
-const Undo = ({
-  pathStep,
-  ctx,
-  canvas,
-  pathPic,
-  setPathMode,
-  setPathStep,
-  pathHistory,
-  saveHistory
-}: ChangedoProps) => {
+const Undo = ({ pathStep, ctx, canvas, pathPic, setPathMode, setPathStep, pathHistory }: ChangedoProps) => {
   if (pathStep === 0) {
     ctx?.clearRect(0, 0, canvas.width, canvas.height);
-    saveHistory();
+    // saveHistory();
   } else {
     pathPic.src = pathHistory[pathStep - 1];
     pathPic.onload = () => {
